Change reader to perform multiple reads
Updated reader to perform multiple reads if the memory size required
is bigger than what the system has. This currently writes to new .wav
files every time but could be appended to previous files. Maybe a good
idea to keep the files separate with the small max file size of a .wav
file